feat(teams,docs,automations,artifacts): Linear anatomy for the B6 areas
Restructured, not reskinned:
- automations: AutomationClient 372 → 131 lines. PageWrapper → PageHeader,
  settings split into AutomationSettingsForm on the shared
  SettingsSection/SettingsToggleRow/SettingsField primitives. The raw
  <button> enabled toggle became a Switch — same behaviour, honest
  affordance. SettingToggle.tsx deleted (verified unreferenced).
- testing-arena: the detail route 429 → 131 lines across four files
  (ReportCard, RunListItem, CodeTestingContent extracted).
- teams: card grid → dense list; TeamMembersTab and TeamReposTab wrapped in
  SettingsSection, TeamReposTab 248 → 116 with TeamAddRepoDialog extracted.
- docs: the three side panels' hand-rolled headers → the PageHeader trio;
  editor prose moved off `prose-sm max-w-none` to `prose` at a 65ch measure.
- analytics/StatusChip: filled pastel pill → tone glyph + neutral text, with
  the name and public API preserved so no call site changed.

Severity dots now go through StatusDot. StatusDot gains a `critical` tone,
and each route maps severity → StatusTone in its own `_utils.ts` instead of
holding raw `bg-*` strings. `high` and `medium` borrow the workflow ramp's
orange and yellow: those are the app's only two mid-warm steps, and adding
severity-specific ones would invent a new tone step, which the colour
ladder forbids. The names read oddly; the red → orange → yellow → grey ramp
does not. The two maps stay duplicated per route on purpose — sharing four
lines would couple two unrelated features.

Real bug fixed: RunListItem had no branch for `pending`, which
evaluationStatusValidator includes, so a queued report rendered an empty
bordered row. Folded into the running branch, matching the existing
aria-label.

Type fix: TeamReposTab's onAdd handler returned the mutation's Promise<null>
into a Promise<void> prop. Now awaited and discarded, which is what the
dialog actually wants.

Worth a look: "Members" and "Codebases" section titles are new; docs panel
headers grew 36-40px → the standard 48px; TeamsClient became a dense list
while ArtifactList kept its card grid deliberately (artifacts are visual,
teams are not).

Left alone with reasons: RunAccordion still nests Buttons inside its
Button trigger (pre-existing invalid HTML; fixing it changes focus
behaviour), ReportCard still renders nothing for `pending`, and the three
iframe white backgrounds are pragma'd — they host untrusted agent HTML.
