+++ apps/web/src/lib/components/analytics/StatusChip.tsx
@@ -4,17 +4,28 @@ import { type Icon as TablerIcon } from "@tabler/icons-react";
 
 type StatusTone = "top" | "positive" | "warning" | "risk" | "neutral";
 
-const toneClass: Record<StatusTone, string> = {
-  top: "bg-warning/15 text-warning",
-  positive: "bg-success/15 text-success",
-  warning: "bg-status-progress-subtle text-status-progress",
-  risk: "bg-destructive/10 text-destructive",
-  neutral: "bg-muted text-muted-foreground",
+// Plain Records (not a computed template string) so Tailwind's static-class
+// scanner can see every literal class here — see StatusDot for the same note.
+const GLYPH_TEXT_CLASS: Record<StatusTone, string> = {
+  top: "text-warning",
+  positive: "text-success",
+  warning: "text-status-progress-bar",
+  risk: "text-destructive",
+  neutral: "text-muted-foreground",
+};
+
+const GLYPH_DOT_CLASS: Record<StatusTone, string> = {
+  top: "bg-warning",
+  positive: "bg-success",
+  warning: "bg-status-progress-bar",
+  risk: "bg-destructive",
+  neutral: "bg-muted-foreground",
 };
 
 /**
- * Soft pastel status pill (the "Top" / "Check" / "Positive" chips from the
- * reference). Background + saturated text from the shared semantic tokens.
+ * Quiet status glyph: a tone-coloured icon (or a small tone-coloured dot when
+ * no icon is given) next to neutral text — never a filled pastel pill, so
+ * status does not compete visually with the content around it.
  */
 export function StatusChip({
   tone,
@@ -26,13 +37,15 @@ export function StatusChip({
   children: ReactNode;
 }) {
   return (
-    <span
-      className={cn(
-        "inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium",
-        toneClass[tone],
+    <span className="inline-flex items-center gap-1 text-2xs font-medium text-muted-foreground">
+      {Icon ? (
+        <Icon size={12} className={GLYPH_TEXT_CLASS[tone]} />
+      ) : (
+        <span
+          aria-hidden
+          className={cn("size-1.5 shrink-0 rounded-full", GLYPH_DOT_CLASS[tone])}
+        />
       )}
-    >
-      {Icon && <Icon size={12} />}
       {children}
     </span>
   );
