+++ apps/web/src/lib/components/artifacts/ArtifactViewer.tsx
@@ -4,7 +4,7 @@ import { Suspense, use, type ReactNode } from "react";
 import { Link } from "@tanstack/react-router";
 import { useQuery } from "convex-helpers/react/cache/hooks";
 import { api } from "@eva/backend";
-import { Spinner } from "@eva/ui";
+import { Button, PageHeader, PageHeaderActions, PageHeaderTitle, Spinner } from "@eva/ui";
 import { IconArrowLeft, IconExternalLink } from "@tabler/icons-react";
 import { ArtifactFrame } from "./ArtifactFrame";
 import { EntityNotFound } from "@/lib/components/EntityNotFound";
@@ -68,53 +68,59 @@ export function ArtifactViewer({ artifactId }: { artifactId: string }) {
   const url = artifact.url;
 
   return (
-    <div className="flex min-h-0 flex-1 flex-col gap-3">
-      <div className="flex shrink-0 items-center gap-3">
-        <Link
-          to="/artifacts"
-          className="flex items-center gap-1 text-sm text-muted-foreground hover:text-foreground"
-        >
-          <IconArrowLeft size={16} />
-          Artifacts
-        </Link>
-        <span className="text-muted-foreground">/</span>
-        <h1 className="min-w-0 truncate text-balance font-medium text-foreground">
-          {artifact.name}
-        </h1>
-        <button
-          type="button"
-          onClick={() =>
-            window.open(`/artifacts/${artifact._id}`, "_blank", "noopener")
-          }
-          className="ml-auto flex shrink-0 items-center gap-1 text-sm text-muted-foreground hover:text-foreground"
-        >
-          <IconExternalLink size={16} />
-          Open in new tab
-        </button>
-      </div>
-      <div className="min-h-0 w-full flex-1 overflow-hidden rounded-surface border border-border bg-white">
-        {url ? (
-          <Suspense
-            fallback={
-              <Centered>
-                <div className="flex flex-col items-center gap-2">
-                  <Spinner />
-                  <span className="text-sm text-muted-foreground">
-                    Loading dashboard…
-                  </span>
-                </div>
-              </Centered>
+    <div className="flex min-h-0 flex-1 flex-col">
+      <PageHeader>
+        <div className="flex min-w-0 items-center gap-2">
+          <Link
+            to="/artifacts"
+            className="flex shrink-0 items-center gap-1 text-2sm text-muted-foreground transition-colors hover:text-foreground"
+          >
+            <IconArrowLeft size={16} />
+            Artifacts
+          </Link>
+          <span className="shrink-0 text-muted-foreground">/</span>
+          <PageHeaderTitle>{artifact.name}</PageHeaderTitle>
+        </div>
+        <PageHeaderActions>
+          <Button
+            type="button"
+            variant="ghost"
+            size="sm"
+            onClick={() =>
+              window.open(`/artifacts/${artifact._id}`, "_blank", "noopener")
             }
           >
-            <ArtifactHtmlBody key={url} url={url} title={artifact.name} />
-          </Suspense>
-        ) : (
-          <Centered>
-            <p className="text-sm text-muted-foreground">
-              Artifact has no content URL.
-            </p>
-          </Centered>
-        )}
+            <IconExternalLink size={16} />
+            Open in new tab
+          </Button>
+        </PageHeaderActions>
+      </PageHeader>
+      <div className="flex min-h-0 flex-1 flex-col p-4">
+        {/* design-check-ignore-next-line — untrusted agent HTML renders on a fixed white iframe plate, not a themed app surface */}
+        <div className="min-h-0 w-full flex-1 overflow-hidden rounded-surface border border-border bg-white">
+          {url ? (
+            <Suspense
+              fallback={
+                <Centered>
+                  <div className="flex flex-col items-center gap-2">
+                    <Spinner />
+                    <span className="text-sm text-muted-foreground">
+                      Loading dashboard…
+                    </span>
+                  </div>
+                </Centered>
+              }
+            >
+              <ArtifactHtmlBody key={url} url={url} title={artifact.name} />
+            </Suspense>
+          ) : (
+            <Centered>
+              <p className="text-sm text-muted-foreground">
+                Artifact has no content URL.
+              </p>
+            </Centered>
+          )}
+        </div>
       </div>
     </div>
   );
