+++ apps/web/src/lib/components/docs/_components/DocHistoryPanel.tsx
@@ -42,39 +42,40 @@ export function DocHistoryPanel({
 
       <div className="scrollbar scroll-fade flex-1 overflow-y-auto">
         {versions.length === 0 && (
-          <p className="px-3 py-6 text-center text-sm text-muted-foreground">
+          <p className="px-3 py-6 text-center text-2sm text-muted-foreground">
             {isRecap
               ? "No prior recap versions yet. Versions are saved when the recap updates on a new push."
               : "No versions saved yet. Versions are created automatically when you stop editing for a few minutes."}
           </p>
         )}
         {versions.map((ver) => (
-          <button
+          <Button
             key={ver._id}
             type="button"
+            variant="ghost"
             onClick={() =>
               onSelectVersion(ver._id === selectedVersionId ? null : ver._id)
             }
             className={cn(
-              "group w-full border-b border-border px-3 py-2.5 text-left transition-colors hover:bg-accent/50",
+              "h-auto w-full flex-col items-stretch justify-start gap-0 whitespace-normal rounded-none border-b border-border px-3 py-2.5 text-left font-normal hover:bg-accent/50",
               ver._id === selectedVersionId && "bg-accent",
             )}
           >
             <div className="flex items-center justify-between gap-2">
-              <MarqueeOnHover className="min-w-0 text-sm font-medium">
+              <MarqueeOnHover className="min-w-0 text-2sm font-medium">
                 {ver.title}
               </MarqueeOnHover>
               {isRecap && ver.headSha ? (
-                <span className="shrink-0 font-mono text-[10px] text-muted-foreground">
+                <span className="shrink-0 font-mono text-3xs text-muted-foreground">
                   {ver.headSha.slice(0, 7)}
                 </span>
               ) : null}
             </div>
             <RelativeDateTime
               at={ver.createdAt}
-              className="text-[10px] text-muted-foreground"
+              className="text-3xs text-muted-foreground"
             />
-          </button>
+          </Button>
         ))}
       </div>
     </div>
