+++ apps/web/src/lib/components/docs/_components/DocVersionDiff.tsx
@@ -53,20 +53,14 @@ export function DocVersionDiff({
         {changes.map((part, i) => {
           if (part.added) {
             return (
-              <ins
-                key={i}
-                className="bg-green-500/20 text-green-700 dark:text-green-400 no-underline"
-              >
+              <ins key={i} className="bg-success/15 text-success no-underline">
                 {part.value}
               </ins>
             );
           }
           if (part.removed) {
             return (
-              <del
-                key={i}
-                className="bg-red-500/20 text-red-700 dark:text-red-400"
-              >
+              <del key={i} className="bg-destructive/15 text-destructive">
                 {part.value}
               </del>
             );
