+++ apps/web/src/routes/_global/teams/_components/TeamMembersTab.tsx
@@ -5,8 +5,7 @@ import { api } from "@eva/backend";
 import type { Id } from "@eva/backend";
 import type { FunctionReturnType } from "convex/server";
 import {
-  Card,
-  CardContent,
+  Badge,
   Button,
   Input,
   Dialog,
@@ -23,6 +22,7 @@ import {
 } from "@eva/ui";
 import { IconTrash, IconUserPlus, IconUsers } from "@tabler/icons-react";
 import { UserInitials } from "@eva/shared";
+import { SettingsSection } from "@/lib/components/settings/SettingsSection";
 import { SettingsEmptyState } from "@/lib/components/settings/SettingsEmptyState";
 
 type Member = FunctionReturnType<typeof api.teamMembers.list>[number];
@@ -111,9 +111,11 @@ export function TeamMembersTab({
   };
 
   return (
-    <>
-      <div className="mb-4 flex justify-end">
-        {isOwner && (
+    <SettingsSection
+      title="Members"
+      description="People with access to this team."
+      action={
+        isOwner ? (
           <Dialog open={dialog.open} onOpenChange={handleDialogChange}>
             <DialogTrigger asChild>
               <Button size="sm">
@@ -165,73 +167,72 @@ export function TeamMembersTab({
               </DialogFooter>
             </DialogContent>
           </Dialog>
-        )}
-      </div>
-      <div className="space-y-2">
-        {members.length === 0 ? (
-          <div className="rounded-surface border border-border bg-card">
-            <SettingsEmptyState
-              icon={IconUsers}
-              title="No members yet"
-              description="Invite teammates by email to collaborate."
-            />
-          </div>
-        ) : (
-          members.map((member) => (
-            <Card key={member._id}>
-              <CardContent className="flex items-center justify-between gap-2 p-3 sm:p-4">
-                <div className="flex min-w-0 items-center gap-2 sm:gap-3">
-                  <UserInitials userId={member.userId} hideLastSeen size="md" />
-                  <div className="min-w-0">
-                    <p data-pii className="truncate text-sm font-medium">
-                      {member.user?.fullName || member.user?.email || "Unknown"}
-                    </p>
-                    <p
-                      data-pii
-                      className="truncate text-xs text-muted-foreground"
-                    >
-                      {member.user?.email}
-                    </p>
-                  </div>
+        ) : undefined
+      }
+      bodyVariant="list"
+    >
+      {members.length === 0 ? (
+        <SettingsEmptyState
+          icon={IconUsers}
+          title="No members yet"
+          description="Invite teammates by email to collaborate."
+        />
+      ) : (
+        <div className="divide-y divide-border">
+          {members.map((member) => (
+            <div
+              key={member._id}
+              className="flex items-center justify-between gap-2 px-4 py-3 transition-colors hover:bg-muted/40"
+            >
+              <div className="flex min-w-0 items-center gap-2 sm:gap-3">
+                <UserInitials userId={member.userId} hideLastSeen size="md" />
+                <div className="min-w-0">
+                  <p data-pii className="truncate text-2sm font-medium">
+                    {member.user?.fullName || member.user?.email || "Unknown"}
+                  </p>
+                  <p
+                    data-pii
+                    className="truncate text-2xs text-muted-foreground"
+                  >
+                    {member.user?.email}
+                  </p>
                 </div>
-                <div className="flex shrink-0 items-center gap-2">
-                  {isOwner && member.userId !== currentUserId ? (
-                    <Select
-                      value={member.role}
-                      onValueChange={(role: "owner" | "member") =>
-                        updateRole({ teamId, userId: member.userId, role })
-                      }
-                    >
-                      <SelectTrigger className="h-7 w-[100px] border-0 bg-secondary text-xs">
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="owner">owner</SelectItem>
-                        <SelectItem value="member">member</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  ) : (
-                    <span className="rounded-full bg-secondary px-2 py-1 text-xs">
-                      {member.role}
-                    </span>
-                  )}
-                  {isOwner && member.userId !== currentUserId && (
-                    <Button
-                      size="icon"
-                      variant="ghost"
-                      onClick={() =>
-                        removeMember({ teamId, userId: member.userId })
-                      }
-                    >
-                      <IconTrash size={14} />
-                    </Button>
-                  )}
-                </div>
-              </CardContent>
-            </Card>
-          ))
-        )}
-      </div>
-    </>
+              </div>
+              <div className="flex shrink-0 items-center gap-2">
+                {isOwner && member.userId !== currentUserId ? (
+                  <Select
+                    value={member.role}
+                    onValueChange={(role: "owner" | "member") =>
+                      updateRole({ teamId, userId: member.userId, role })
+                    }
+                  >
+                    <SelectTrigger className="h-7 w-[100px] border-0 bg-secondary text-xs">
+                      <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                      <SelectItem value="owner">owner</SelectItem>
+                      <SelectItem value="member">member</SelectItem>
+                    </SelectContent>
+                  </Select>
+                ) : (
+                  <Badge variant="quiet">{member.role}</Badge>
+                )}
+                {isOwner && member.userId !== currentUserId && (
+                  <Button
+                    size="icon-sm"
+                    variant="ghost"
+                    onClick={() =>
+                      removeMember({ teamId, userId: member.userId })
+                    }
+                  >
+                    <IconTrash size={14} />
+                  </Button>
+                )}
+              </div>
+            </div>
+          ))}
+        </div>
+      )}
+    </SettingsSection>
   );
 }
