+++ apps/web/src/routes/_global/teams/_components/TeamReposTab.tsx
@@ -1,25 +1,11 @@
-import { useState } from "react";
 import { useMutation } from "convex/react";
 import { api } from "@eva/backend";
 import type { Id } from "@eva/backend";
 import type { FunctionReturnType } from "convex/server";
-import {
-  Button,
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-  DialogFooter,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@eva/ui";
-import { IconPlus, IconFolder } from "@tabler/icons-react";
+import { IconFolder } from "@tabler/icons-react";
 import { TeamRepoCard } from "./TeamRepoCard";
+import { TeamAddRepoDialog } from "./TeamAddRepoDialog";
+import { SettingsSection } from "@/lib/components/settings/SettingsSection";
 import { SettingsEmptyState } from "@/lib/components/settings/SettingsEmptyState";
 
 type Repo = FunctionReturnType<typeof api.githubRepos.listByTeam>[number];
@@ -90,149 +76,35 @@ export function TeamReposTab({
     }
   });
 
-  const [dialog, setDialog] = useState({
-    open: false,
-    selectedRepoId: "",
-    error: "",
-    isSubmitting: false,
-  });
-
-  const handleDialogChange = (open: boolean) => {
-    if (!open) {
-      setDialog({
-        open: false,
-        selectedRepoId: "",
-        error: "",
-        isSubmitting: false,
-      });
-    } else {
-      setDialog((prev) => ({ ...prev, open: true }));
-    }
-  };
-
-  const handleAddRepo = async () => {
-    if (!dialog.selectedRepoId) {
-      setDialog((prev) => ({
-        ...prev,
-        error: "Please select a repository",
-      }));
-      return;
-    }
-
-    const foundRepo = allRepos.find((r) => r._id === dialog.selectedRepoId);
-    if (!foundRepo) {
-      setDialog((prev) => ({ ...prev, error: "Repository not found" }));
-      return;
-    }
-
-    setDialog((prev) => ({ ...prev, error: "", isSubmitting: true }));
-
-    try {
-      await assignRepo({ teamId, repoId: foundRepo._id });
-      setDialog({
-        open: false,
-        selectedRepoId: "",
-        error: "",
-        isSubmitting: false,
-      });
-    } catch (err) {
-      const errorMessage =
-        err instanceof Error ? err.message : "Failed to add repository";
-      setDialog((prev) => ({
-        ...prev,
-        error: errorMessage,
-        isSubmitting: false,
-      }));
-    }
-  };
-
   const availableRepos = allRepos.filter((r) => r.teamId !== teamId);
 
   return (
-    <>
-      <div className="mb-4 flex justify-end">
-        {isOwner && (
-          <Dialog open={dialog.open} onOpenChange={handleDialogChange}>
-            <DialogTrigger asChild>
-              <Button size="sm">
-                <IconPlus size={16} className="mr-1.5" />
-                Add Repository
-              </Button>
-            </DialogTrigger>
-            <DialogContent>
-              <DialogHeader>
-                <DialogTitle>Add Repository</DialogTitle>
-                <DialogDescription>
-                  Assign a codebase to this team.
-                </DialogDescription>
-              </DialogHeader>
-              <div className="space-y-4">
-                {availableRepos.length === 0 ? (
-                  <p className="text-sm text-muted-foreground">
-                    No unassigned codebases available.
-                  </p>
-                ) : (
-                  <Select
-                    value={dialog.selectedRepoId}
-                    onValueChange={(value) =>
-                      setDialog((prev) => ({
-                        ...prev,
-                        selectedRepoId: value,
-                        error: "",
-                      }))
-                    }
-                    disabled={dialog.isSubmitting}
-                  >
-                    <SelectTrigger>
-                      <SelectValue placeholder="Select a repository" />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {availableRepos.map((repo) => (
-                        <SelectItem key={repo._id} value={repo._id}>
-                          {repo.owner}/{repo.name}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                )}
-                {dialog.error ? (
-                  <div className="rounded-surface border border-destructive/50 bg-destructive/10 p-3">
-                    <p className="text-sm text-destructive">{dialog.error}</p>
-                  </div>
-                ) : null}
-              </div>
-              <DialogFooter>
-                <Button
-                  variant="secondary"
-                  onClick={() => handleDialogChange(false)}
-                  disabled={dialog.isSubmitting}
-                >
-                  Cancel
-                </Button>
-                <Button
-                  onClick={handleAddRepo}
-                  disabled={
-                    dialog.isSubmitting || availableRepos.length === 0
-                  }
-                >
-                  {dialog.isSubmitting ? "Adding…" : "Add"}
-                </Button>
-              </DialogFooter>
-            </DialogContent>
-          </Dialog>
-        )}
-      </div>
-      <div className="space-y-2">
-        {repos.length === 0 ? (
-          <div className="rounded-surface border border-border bg-card">
-            <SettingsEmptyState
-              icon={IconFolder}
-              title="No codebases yet"
-              description="Assign a repository to this team."
-            />
-          </div>
-        ) : (
-          repos.map((repo) => (
+    <SettingsSection
+      title="Codebases"
+      description="Repositories assigned to this team."
+      action={
+        isOwner ? (
+          <TeamAddRepoDialog
+            availableRepos={availableRepos}
+            onAdd={async (repoId) => {
+              // Awaited and discarded: the mutation resolves to null, and the
+              // dialog only needs to know when it settled.
+              await assignRepo({ teamId, repoId });
+            }}
+          />
+        ) : undefined
+      }
+      bodyVariant="list"
+    >
+      {repos.length === 0 ? (
+        <SettingsEmptyState
+          icon={IconFolder}
+          title="No codebases yet"
+          description="Assign a repository to this team."
+        />
+      ) : (
+        <div className="divide-y divide-border">
+          {repos.map((repo) => (
             <TeamRepoCard
               key={repo._id}
               repo={repo}
@@ -240,9 +112,9 @@ export function TeamReposTab({
               isOwner={isOwner}
               onRemove={(repoId) => removeRepo({ teamId, repoId })}
             />
-          ))
-        )}
-      </div>
-    </>
+          ))}
+        </div>
+      )}
+    </SettingsSection>
   );
 }
